# specs/001-python-llm-agent/contracts/api.yaml
openapi: 3.0.3
info:
  title: 多Agent加密货币量化交易分析系统 API
  description: |
    基于Python的多Agent虚拟货币量化交易分析系统API文档

    系统特性：
    - 5个专业agent（新闻收集、做多分析、做空分析、策略生成、交易执行）
    - 支持前5大主流交易所
    - 实时策略分析和自动交易执行
    - 动态资金管理和风险控制
    - 移动端实时监控支持

  version: 1.0.0
  contact:
    name: API Support
    email: support@crypto-ai-trading.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.crypto-ai-trading.com/v1
    description: 生产环境
  - url: https://staging-api.crypto-ai-trading.com/v1
    description: 测试环境
  - url: http://localhost:8000/v1
    description: 开发环境
paths:
...
  /news:
    get:
      tags:
        - 新闻数据
      summary: 获取新闻列表
      description: 获取最新的加密货币相关新闻
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
          description: 返回条数
        - name: sentiment
          in: query
          schema:
            type: string
            enum: [positive, negative, neutral]
          description: 情感过滤
        - name: symbol
          in: query
          schema:
            type: string
          description: 相关符号过滤
      responses:
        '200':
          description: 新闻列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  news:
                    type: array
                    items:
                      $ref: '#/components/schemas/NewsItem'
  /news/summary:
    get:
      tags:
        - 新闻数据
      summary: 获取新闻摘要
      description: 获取AI生成的新闻摘要和市场影响分析
      security:
        - bearerAuth: []
      parameters:
        - name: hours
          in: query
          schema:
            type: integer
            default: 24
          description: 摘要时间范围（小时）
      responses:
        '200':
          description: 新闻摘要
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsSummary'

  # ==================== 技术分析 ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    NewsItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        url:
          type: string
        source:
          type: string
        published_at:
          type: string
          format: date-time
        relevance_score:
          type: number
          format: decimal
          minimum: 0
          maximum: 1
        sentiment:
          type: string
          enum: [positive, negative, neutral]
        related_symbols:
          type: array
          items:
            type: string
    NewsSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        summary_text:
          type: string
        key_points:
          type: array
          items:
            type: string
        market_impact:
          type: string
          enum: [high, medium, low]
        news_count:
          type: integer
        time_period_hours:
          type: integer
        confidence_score:
          type: number
          format: decimal
        created_at:
          type: string
          format: date-time

    # ==================== 技术分析模型 ====================
...